backend/integration-test/requirements.txt
requests==2.18.4
psycopg2-binary==2.7.4
